implement policy to prevent change of order after it's send or paid

+++ app/Http/Controllers/OrderItemController.php
@@ -88,6 +88,8 @@ class OrderItemController extends Controller
      */
     public function destroy(OrderItem $orderItem)
     {
+        $this->authorize('modify', $orderItem);
+
         $orderItem->delete();
     }
 }
